Include unit tests for Python code
The tests are run by tox using the pytest testing framework. A basic
`setup.py` file is included to allow tox to build a source distribution
from this project.

A GitHub action is added to automatically run the unit tests when code
is pushed to the repository and a PR is created or updated. It is also
scheduled to run every Friday at 18:00 UTC.