Include end-to-end tests for JavaScript code
The tests are run by tox using the pytest testing framework combined
with the Selenium WebDriver automation framework.

A GitHub action is added to automatically run the tests at BrowserStack
when code is pushed to the repository and a PR is created or updated. It
is also scheduled to run every Friday at 18:00 UTC.