Add JavaScript and Python code linters
Linting JavaScript code is done by ESLint, the Python code is linted by
Flake8. Some Python code is updated to fix linting errors.

A GitHub action is added to automatically run the linters when code is
pushed to the repository and when a PR is created or updated. Since
linting is only relevant upon code changes, it's not run following a
time schedule.