Fix to ensure added forms can be deleted
For `<template>` elements, the `template.content` property returns a
DocumentFragment instead of a Node. These can be added easily to the
DOM using `appendChild`, but deleting them using `removeChild` is
non-trivial in such cases, as pointed out it #6.

To overcome this problem, the `template.content.firstElementChild` is
added to the DOM instead, which is a Node. This does require the
template element to contain only a single element and the library now
also checks for this upon initialization, but in practice this was
already the case.

End-to-end tests are updated to assert this check is performed and to
ensure added forms can indeed be deleted. The documentation is updated
to clarify that each form should be contained inside a single parent
element denoted by the `formSelector` parameter.